requirements.txt
kafka-python==2.0.2
pyspark==3.5.0
psycopg2-binary==2.9.7
pandas==2.1.0
numpy==1.25.2
scikit-learn==1.3.0
mlflow==2.7.1
python-dotenv==1.0.0
schedule==1.2.0
tensorflow==2.15.0
